requirements.txt
google-cloud-vision>=3.4.4
google-auth-oauthlib>=1.0.0
google-auth>=2.22.0
Pillow>=10.0.0
openai>=1.3.0